Create wercker.yml
Let's look at the two sections of YAML that we've just added. The first section describes a box. A box is the image that you want Wercker to pull from a Docker registry (the default is Docker Hub) to build your pipeline from. In our case, we need a Java environment to run our application, so we will pull the openjdk image from Docker Hub.

The second section defines our first pipeline, the build pipeline. Our build pipeline consists of two steps, one to install any packages required for our build process (install-packages), and one to execute a shell command to invoke the build (script).

Since our application is built using Maven, we will instruct Wercker to install the Maven package first, using the install-packages step. Then we'll build the application the same way we would normally, invoking Maven in the script step.