Update wercker.yml
This section of YAML defines our second pipeline, push-release. It uses a type of step we haven't seen before, internal/docker-push. This is a command developed by Wercker, which will take the current pipeline image and push it to a Docker registry. You can, however, use a different repository -- the step supports a repository parameter. You can find more info in the documentation.

There are two more things to note about this pipeline definition. First, the cmd parameter tells Docker how to start our application. Note that this is just the start script generated by Maven, exactly as we would have started it outside of Docker.

Second, we have made use of several environment variables. We are tagging the release with the Git branch name and the Git commit hash so that we can trace our running application version back to a specific Git commit. These two variables (along with others) are automatically made available by Wercker. The $PORT environment variable, on the other hand, references the one that we defined ourselves in the previous step.